requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker-pycreds==0.4.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
pandas
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn
sentry-sdk==2.26.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.3
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.19.9
yfinance
xlrd==2.0.2